requirements.txt
tornado
numpy
opencv-python-headless
onnxruntime
Shapely
pyclipper
Pillow
